refactor: Add modules to a separate folder
