+++ local-modules/getAllTalkers.js
